:construction: Refactoring code, create components ReactCounter and ReduxCounter
